Move layout logic out of `plot/`
